Simplify hextiling implementations
This changelist simplifies the hextiling implementations in GLSL and MDL, with the goal of making them more maintainable and extensible to new languages in the future.

One natural next step will be to implement the equivalent hextiling logic in OSL, using the simplified approach from this changelist as a template.